Implement branch handling and instruction fetching in the pipeline
- Added `BranchResult` structure to encapsulate branch results.
- Removed `MemoryRS` class from `lsb.hpp` as it is no longer needed.
- Updated `RobIDType` to `uint32_t` in `constants.hpp` for simplicity.
- Enhanced `Decoder` class to handle branch instructions and integrate a predictor.
- Introduced `Fetcher` class to manage instruction fetching from a buffer.
- Created `PCLogic` class to manage program counter updates and flush signals.
- Implemented `Predictor` class for branch prediction logic.
- Defined `Instruction` structure to represent decoded instructions.
- Updated `Control` class to handle instruction fetching and branch processing.
- Refactored `RegisterFile` and `ReorderBuffer` classes for better state management.
- Added utility functions for converting integers to byte arrays.
- Created unit test file for logger functionality.

+++ include/constants.hpp
@@ -8,7 +8,7 @@ using RegDataType = uint32_t;
 using WordType = uint32_t;
 
 // 0 is left for unused
-using RobIDType = uint8_t;
+using RobIDType = uint32_t; // for simplicity to avoid considering circular situations
 
 
 using MemAddrType = uint32_t;
